feat(server): modular REST domain routes; BP1-1 docs and architecture fixes
Split registerRestDomainRoutes into per-domain modules under server/routes/ with
restDomainShared and RestRouteContext. Add trilingual BP1-1 order sketch docs,
extend ADR-0009, and align operational flow with canonical states.

Fix architecture (EN/ES/PT): document modular API layout and session-based auth;
remove obsolete claim that the API had no authentication. Master plan BP0-4 now
cites registerRestDomainRoutes and server/routes/.

Verification: npm run check:docs-map, npm run type-check, npm run test (405).

Related to #65, #69, #77.
