+++ docs/DOCUMENT_LOCALE_MAP.md
@@ -34,6 +34,7 @@ This table maps each **logical document** to its path in `docs/en/`, `docs/es/`,
 | RBAC matrix (roles, permissions, scopes) | [quality/rbac-matrix-roles-permissions-scopes.md](en/quality/rbac-matrix-roles-permissions-scopes.md) | [quality/matriz-rbac-roles-permisos-scopes.md](es/quality/matriz-rbac-roles-permisos-scopes.md) | [quality/matriz-rbac-funcoes-permissoes-scopes.md](pt-br/quality/matriz-rbac-funcoes-permissoes-scopes.md) |
 | IAM model (sessions, audit, endpoint auth) | [quality/iam-model-sessions-audit.md](en/quality/iam-model-sessions-audit.md) | [quality/modelo-iam-sesiones-auditoria.md](es/quality/modelo-iam-sesiones-auditoria.md) | [quality/modelo-iam-sessoes-auditoria.md](pt-br/quality/modelo-iam-sessoes-auditoria.md) |
 | Operational flow: order → delivery → collection (design) | [quality/operational-flow-order-delivery-collection.md](en/quality/operational-flow-order-delivery-collection.md) | [quality/flujo-operativo-pedido-entrega-cobranza.md](es/quality/flujo-operativo-pedido-entrega-cobranza.md) | [quality/fluxo-operacional-pedido-entrega-cobranca.md](pt-br/quality/fluxo-operacional-pedido-entrega-cobranca.md) |
+| Order domain — BP1-1 implementation sketch (Prisma/OpenAPI draft) | [quality/order-domain-implementation-sketch.md](en/quality/order-domain-implementation-sketch.md) | [quality/boceto-implementacion-dominio-pedido.md](es/quality/boceto-implementacion-dominio-pedido.md) | [quality/rascunho-implementacao-dominio-pedido.md](pt-br/quality/rascunho-implementacao-dominio-pedido.md) |
 | Specs index | [specs/index.md](en/specs/index.md) | [specs/indice.md](es/specs/indice.md) | [specs/indice.md](pt-br/specs/indice.md) |
 | Technical manual | [specs/technical-manual.md](en/specs/technical-manual.md) | [specs/manual-tecnico.md](es/specs/manual-tecnico.md) | [specs/manual-tecnico.md](pt-br/specs/manual-tecnico.md) |
 | Functional requirements | [specs/functional-requirements.md](en/specs/functional-requirements.md) | [specs/requisitos-funcionales.md](es/specs/requisitos-funcionales.md) | [specs/requisitos-funcionais.md](pt-br/specs/requisitos-funcionais.md) |
