+++ docs/en/quality/master-plan-bizcode-execution.md
@@ -46,7 +46,7 @@ Priorities below are **planning** items; verification is against the repository 
 | BP0-1 | P0 | Trilingual execution package (this doc + RBAC matrix + IAM + order-flow design) | Files exist under `docs/en|es|pt-br/quality/` per [DOCUMENT_LOCALE_MAP.md](../../DOCUMENT_LOCALE_MAP.md); `npm run check:docs-map` passes |
 | BP0-2 | P0 | RBAC source of truth documented | Matrix matches `ROLE_PERMISSIONS` in [`src/lib/rbac.ts`](../../../src/lib/rbac.ts) |
 | BP0-3 | P0 | IAM and sessions documented | Describes `Tenant`, `AppUser`, `AppSession`, `AuditEvent` in [`prisma/schema.prisma`](../../../prisma/schema.prisma) and flow in [`server/auth.ts`](../../../server/auth.ts) |
-| BP0-4 | P0 | Auth on domain API | Routes under `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` use `requirePermission` in [`server/createApp.ts`](../../../server/createApp.ts); contract in [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
+| BP0-4 | P0 | Auth on domain API | Routes under `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` use `requirePermission` when registered from [`server/registerRestDomainRoutes.ts`](../../../server/registerRestDomainRoutes.ts) (per-domain modules under `server/routes/`); app wiring in [`server/createApp.ts`](../../../server/createApp.ts); contract in [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
 | BP1-1 | P1 | Order (`pedido`) domain | **Future:** Prisma model + migration + routes only when specified in OpenAPI and implemented in `server/` |
 | BP1-2 | P1 | E2E / integration coverage for critical paths | Align with [testing-strategy.md](testing-strategy.md) and existing Playwright/Postgres tooling |
 | BP1-3 | P1 | Channel scope enforcement | **Implemented in current codebase:** `requirePermission` in [`server/auth.ts`](../../../server/auth.ts) validates optional `x-bizcode-channel` against `AuthScope.channels`; invalid header returns `400`, out-of-scope returns `403`. Covered by [`tests/server/scope-channel.test.ts`](../../../tests/server/scope-channel.test.ts) |
