+++ docs/es/quality/ejecucion-plan-maestro-bizcode.md
@@ -46,7 +46,7 @@ Las prioridades siguientes son ítems de **planificación**; la verificación es
 | BP0-1 | P0 | Paquete trilingüe de ejecución (este doc + matriz RBAC + IAM + diseño de flujo pedido) | Archivos bajo `docs/en|es|pt-br/quality/` según [DOCUMENT_LOCALE_MAP.md](../../DOCUMENT_LOCALE_MAP.md); `npm run check:docs-map` OK |
 | BP0-2 | P0 | RBAC documentado como fuente de verdad | La matriz coincide con `ROLE_PERMISSIONS` en [`src/lib/rbac.ts`](../../../src/lib/rbac.ts) |
 | BP0-3 | P0 | IAM y sesiones documentados | Describe `Tenant`, `AppUser`, `AppSession`, `AuditEvent` en [`prisma/schema.prisma`](../../../prisma/schema.prisma) y el flujo en [`server/auth.ts`](../../../server/auth.ts) |
-| BP0-4 | P0 | Auth en API de dominio | Rutas bajo `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` usan `requirePermission` en [`server/createApp.ts`](../../../server/createApp.ts); contrato en [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
+| BP0-4 | P0 | Auth en API de dominio | Rutas bajo `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` usan `requirePermission` al registrarse desde [`server/registerRestDomainRoutes.ts`](../../../server/registerRestDomainRoutes.ts) (módulos en `server/routes/`); cableado de la app en [`server/createApp.ts`](../../../server/createApp.ts); contrato en [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
 | BP1-1 | P1 | Dominio pedido (`pedido`) | **Futuro:** modelo Prisma + migración + rutas solo cuando consten en OpenAPI y en `server/` |
 | BP1-2 | P1 | Cobertura E2E / integración de caminos críticos | Alineado con [estrategia-pruebas.md](estrategia-pruebas.md) y herramientas Playwright/Postgres existentes |
 | BP1-3 | P1 | Refuerzo de alcance por canal | **Implementado en el código actual:** `requirePermission` en [`server/auth.ts`](../../../server/auth.ts) valida `x-bizcode-channel` opcional contra `AuthScope.channels`; cabecera inválida devuelve `400` y canal fuera de alcance devuelve `403`. Cubierto por [`tests/server/scope-channel.test.ts`](../../../tests/server/scope-channel.test.ts) |
