+++ docs/pt-br/quality/execucao-plano-mestre-bizcode.md
@@ -46,7 +46,7 @@ As prioridades abaixo são itens de **planejamento**; a verificação é contra 
 | BP0-1 | P0 | Pacote trilíngue de execução (este doc + matriz RBAC + IAM + desenho de fluxo pedido) | Arquivos em `docs/en|es|pt-br/quality/` conforme [DOCUMENT_LOCALE_MAP.md](../../DOCUMENT_LOCALE_MAP.md); `npm run check:docs-map` OK |
 | BP0-2 | P0 | RBAC documentado como fonte da verdade | A matriz coincide com `ROLE_PERMISSIONS` em [`src/lib/rbac.ts`](../../../src/lib/rbac.ts) |
 | BP0-3 | P0 | IAM e sessões documentados | Descreve `Tenant`, `AppUser`, `AppSession`, `AuditEvent` em [`prisma/schema.prisma`](../../../prisma/schema.prisma) e o fluxo em [`server/auth.ts`](../../../server/auth.ts) |
-| BP0-4 | P0 | Auth na API de domínio | Rotas sob `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` usam `requirePermission` em [`server/createApp.ts`](../../../server/createApp.ts); contrato em [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
+| BP0-4 | P0 | Auth na API de domínio | Rotas sob `/api/clientes`, `/api/articulos`, `/api/rubros`, `/api/facturas`, `/api/formas-pago` usam `requirePermission` ao serem registradas em [`server/registerRestDomainRoutes.ts`](../../../server/registerRestDomainRoutes.ts) (módulos em `server/routes/`); composição da app em [`server/createApp.ts`](../../../server/createApp.ts); contrato em [`docs/api/openapi.yaml`](../../api/openapi.yaml) |
 | BP1-1 | P1 | Domínio pedido (`pedido`) | **Futuro:** modelo Prisma + migração + rotas apenas quando constarem no OpenAPI e em `server/` |
 | BP1-2 | P1 | Cobertura E2E / integração de fluxos críticos | Alinhado a [estrategia-testes.md](estrategia-testes.md) e ferramentas Playwright/Postgres existentes |
 | BP1-3 | P1 | Enforcement de escopo por canal | **Implementado no código atual:** `requirePermission` em [`server/auth.ts`](../../../server/auth.ts) valida `x-bizcode-channel` opcional contra `AuthScope.channels`; cabeçalho inválido retorna `400` e canal fora do escopo retorna `403`. Coberto por [`tests/server/scope-channel.test.ts`](../../../tests/server/scope-channel.test.ts) |
